+++ .pre-commit-config.yaml
@@ -18,7 +18,7 @@ exclude: "^\
   (.vscode/.*)|\
   (^tests)|\
   (docs/api/.*)|\
-  (Makefile)
+  (^Makefile)
   "
 
 ci:
@@ -47,7 +47,7 @@ repos:
     hooks:
       - id: prettier
         files: \.(html|json|markdown|md|yaml|yml)$
-        exclude: (^docs/api/.*)
+        exclude: (^docs/api/.* | Makefile)
 
   # yapf formatter
   - repo: https://github.com/pre-commit/mirrors-yapf
@@ -62,13 +62,6 @@ repos:
     rev: 5.10.1
     hooks:
       - id: isort
-
-  # Changes tabs to spaces
-  - repo: https://github.com/Lucas-C/pre-commit-hooks
-    rev: v1.1.13
-    hooks:
-      - id: remove-tabs
-        exclude: ^(docs)
 
   # CMake formatting
   - repo: https://github.com/cheshirekow/cmake-format-precommit
